tests improvements
