+++ conans/test/functional/old/sysroot_test.py
@@ -2,7 +2,6 @@ import os
 import unittest
 
 from conans.test.utils.tools import TestClient
-from conans.util.files import load
 
 
 class SysrootTest(unittest.TestCase):
